Move Template.{extensions,copy_without_render} to Variables
