+++ src/cutty/core/render.py
@@ -11,7 +11,7 @@ from .template import Template
 
 def create_environment(template: Template) -> jinja2.Environment:
     """Create the Jinja environment."""
-    extensions_ = extensions.load(extra=template.extensions)
+    extensions_ = extensions.load(extra=template.variables.extensions)
     return jinja2.Environment(  # noqa: S701
         loader=jinja2.FileSystemLoader(str(template.repository)),
         extensions=extensions_,
